Refine - 增加誤刪的警示，Refine Product add/edit/destroy UX

+++ app/controllers/admin/events_controller.rb
@@ -1,13 +1,10 @@
 class Admin::EventsController < AdminController
-
-  expose :events, -> { Event.all.order(created_at: :desc).paginate(:page => params[:page], per_page: 30) }
+  expose :events, -> { Event.all.order(created_at: :desc).paginate(page: params[:page], per_page: 30) }
   expose :event
 
-  def index
-  end
+  def index; end
 
-  def show
-  end
+  def show; end
 
   def new
     event.build_meta
@@ -21,8 +18,7 @@ class Admin::EventsController < AdminController
     end
   end
 
-  def edit
-  end
+  def edit; end
 
   def update
     if event.update event_params
@@ -33,14 +29,19 @@ class Admin::EventsController < AdminController
   end
 
   def destroy
-    event.destroy
-    redirect_to admin_events_path, notice: t('helpers.successfully_destroy')
+    if event.enable
+      redirect_to admin_events_path, alert: t('helpers.destroy_fail')
+    else
+      event.destroy
+      redirect_to admin_events_path, notice: t('helpers.successfully_destroy')
+    end
   end
 
   private
+
   def event_params
     params.require(:event)
           .permit(:title, :desc, :link_name, :enable,
-                    meta_attributes: [:id, :event_id, :desc, :keywords, :image] )
+                  meta_attributes: [:id, :event_id, :desc, :keywords, :image])
   end
 end
